Avoid implicitly promoting `float` to `double`
CURL reports current position and total size as a pair of `double`s, and
it makes little sense to store those as `float`s if we end up converting
them to `double`s again (for comparisons, and to calculate the progress).

+++ include/download.h
@@ -53,8 +53,8 @@ private:
 	std::string fn;
 	std::string url_;
 	DlStatus download_status;
-	float cursize;
-	float totalsize;
+	double cursize;
+	double totalsize;
 	double curkbps;
 	unsigned long offs;
 	PbController* ctrl;
